Allowing acceptances across multiple offers

+++ db/migrate/20120815161110_create_price_intervals.rb
@@ -3,7 +3,7 @@ class CreatePriceIntervals < ActiveRecord::Migration
     create_table :price_intervals do |t|
       t.datetime :start_time
       t.datetime :end_time
-      t.double :price_per_hour
+      t.float :price_per_hour
       t.int :price_plan_id
 
       t.timestamps
